+++ tests/ops/test_qubit_ops.py
@@ -739,6 +739,35 @@ class TestOperations:
 
         assert np.allclose(decomposed_matrix, global_phase * op.matrix, atol=tol, rtol=0)
 
+    @pytest.mark.parametrize(
+        "op",
+        [
+            (qml.Hadamard(wires=0)),
+            (qml.PauliX(wires=0)),
+            (qml.PauliY(wires=0)),
+            (qml.PauliZ(wires=0)),
+            (qml.S(wires=0)),
+            (qml.T(wires=0)),
+            (qml.SX(wires=0)),
+            (qml.RX(0.3, wires=0)),
+            (qml.RY(0.3, wires=0)),
+            (qml.RZ(0.3, wires=0)),
+            (qml.PhaseShift(0.3, wires=0)),
+            (qml.Rot(0.3, 0.4, 0.5, wires=0)),
+        ],
+    )
+    def test_single_qubit_rot_angles(self, op):
+        """Tests that the Rot gates yielded by single_qubit_rot_angles
+        are equivalent to the true operations up to a global phase."""
+        angles = op.single_qubit_rot_angles()
+        obtained_mat = qml.Rot(*angles, wires=0).matrix
+
+        # Check whether the two matrices are each others conjugate transposes
+        mat_product = qml.math.dot(op.matrix, qml.math.conj(obtained_mat.T))
+        mat_product /= mat_product[0, 0]
+
+        assert qml.math.allclose(mat_product, I)
+
     def test_CY_decomposition(self, tol):
         """Tests that the decomposition of the CY gate is correct"""
         op = qml.CY(wires=[0, 1])